models(transformer-perf): address review — docs, overflow-safe guard, honest bench (#25)
Review follow-ups on the ForwardInto overload:

- ForwardInto/Forward XML docs no longer claim Forward performs a final memcpy;
  both entry points now have the LM head write straight into the destination,
  and the documented difference is buffer ownership (per-call UnmanagedTensor vs
  caller-provided span). Step-4 inline comment reworded to match.
- Size guard: widen `seqLen * VocabSize` to long before comparing, so large
  context x vocab combinations cannot wrap int and silently disable the check.
  Documented the `ArgumentException` and added a boundary test at
  seqLen x vocabSize - 1.
- Benchmark: rename the old alloc+copy+free case to PreRefactor_AllocCopyFree
  (historical context, no current path does it) and add Forward_AllocOnly as the
  Baseline, which is what Forward actually costs today. ForwardInto is now
  measured against the real baseline instead of an overstated one.

+++ src/DotLLM.Models/Architectures/TransformerModel.cs
@@ -131,6 +131,13 @@ public sealed unsafe class TransformerModel : IModel
     /// <param name="deviceId">Target device for computation.</param>
     /// <param name="kvCache">Optional KV-cache. When null, behaves identically to the uncached forward pass.</param>
     /// <returns>Logits tensor of shape [seqLen, vocab_size] for all input positions.</returns>
+    /// <remarks>
+    /// The LM head writes its output straight into the returned tensor — there is no intermediate
+    /// copy — but the tensor itself is allocated fresh on every call. Callers that own a reusable
+    /// logits buffer should prefer
+    /// <see cref="ForwardInto(ReadOnlySpan{int}, ReadOnlySpan{int}, IKvCache?, Span{float})"/>,
+    /// which avoids that per-call allocation.
+    /// </remarks>
     public ITensor Forward(ReadOnlySpan<int> tokenIds, ReadOnlySpan<int> positions,
                            int deviceId, IKvCache? kvCache)
     {
@@ -142,21 +149,28 @@ public sealed unsafe class TransformerModel : IModel
     }
 
     /// <summary>
-    /// Runs a forward pass and writes the resulting logits directly into <paramref name="logitsOut"/>,
-    /// avoiding the <see cref="UnmanagedTensor"/> allocation and the final memcpy that
-    /// <see cref="Forward(ReadOnlySpan{int}, ReadOnlySpan{int}, int, IKvCache?)"/> performs.
-    /// This is the recommended entry point for tight decode loops that already own a
-    /// reusable logits buffer (typical of <c>TextGenerator</c>'s sampling path).
+    /// Runs a forward pass and writes the resulting logits directly into <paramref name="logitsOut"/>.
+    /// Both this overload and <see cref="Forward(ReadOnlySpan{int}, ReadOnlySpan{int}, int, IKvCache?)"/>
+    /// have the LM head write straight into their destination buffer (neither copies through
+    /// <c>_state.Logits</c> any more); the difference is ownership — <c>Forward</c> allocates a fresh
+    /// <see cref="UnmanagedTensor"/> of [seqLen × vocabSize] floats per call, while this overload reuses
+    /// the caller's buffer and so allocates nothing. This is the recommended entry point for tight
+    /// decode loops that already own a reusable logits buffer (typical of <c>TextGenerator</c>'s
+    /// sampling path).
     /// </summary>
     /// <param name="tokenIds">Input token IDs for this step.</param>
     /// <param name="positions">Position indices for each token.</param>
     /// <param name="kvCache">Optional KV-cache.</param>
     /// <param name="logitsOut">Destination span for logits. Must have length &gt;= seqLen × vocabSize.</param>
+    /// <exception cref="ArgumentException"><paramref name="logitsOut"/> is shorter than seqLen × vocabSize.</exception>
     public void ForwardInto(ReadOnlySpan<int> tokenIds, ReadOnlySpan<int> positions,
                             IKvCache? kvCache, Span<float> logitsOut)
     {
         int seqLen = tokenIds.Length;
-        int needed = seqLen * Config.VocabSize;
+        // Widen before multiplying: seqLen × vocabSize exceeds int.MaxValue for large
+        // context × vocab combinations, and a silently-wrapped (possibly negative)
+        // product would turn this guard into a no-op.
+        long needed = (long)seqLen * Config.VocabSize;
         if (logitsOut.Length < needed)
             throw new ArgumentException(
                 $"logitsOut span is too small: need {needed} floats ({seqLen} × {Config.VocabSize}), got {logitsOut.Length}.",
@@ -449,11 +463,11 @@ public sealed unsafe class TransformerModel : IModel
         }
 
         // 4. LM HEAD — all positions (enables batched speculative decoding verification).
-        //    Writes directly into the caller-provided `logitsDst` buffer, eliminating the
-        //    previous `_state.Logits → result` copy. The `Forward(...)` entry point passes
-        //    a freshly-allocated UnmanagedTensor's pointer; `ForwardInto(...)` passes the
-        //    caller's pinned Span<float>, avoiding the allocation altogether for hot-path
-        //    decode loops.
+        //    Writes directly into the caller-provided `logitsDst` buffer. This replaces the
+        //    old `_state.Logits → result` copy for BOTH entry points: `Forward(...)` passes
+        //    a freshly-allocated UnmanagedTensor's pointer (allocation, no copy), while
+        //    `ForwardInto(...)` passes the caller's pinned Span<float> (no allocation,
+        //    no copy) for hot-path decode loops.
         {
             var rwOutput = _weights.RepackedOutput ?? default;
             GemmInterleaved(_weights.OutputWeight, _weights.OutputQuantType,
